backend/auth-requirements.txt
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
Markdown
Pillow
sendgrid